Remove `compiler` and move `with_backend` to handler.
