+++ ivy/utils/backend/__init__.py
@@ -1,5 +1,4 @@
 # flake8: noqa
 from . import handler
 from .handler import *
-from . import compiler
 from ._importlib import _ivy_import_module
